+++ lib/cipi-api-sudoers.sh
@@ -67,6 +67,8 @@ www-data ALL=(root) NOPASSWD: /usr/local/bin/cipi app create *, \
                                /usr/local/bin/cipi search list *, \
                                /usr/local/bin/cipi search enable *, \
                                /usr/local/bin/cipi search disable *, \
+                               /usr/local/bin/cipi package list, \
+                               /usr/local/bin/cipi package list *, \
                                /usr/local/bin/cipi php list, \
                                /usr/local/bin/cipi php list *, \
                                /usr/local/bin/cipi php install *, \